调整金额相关为整数

+++ src/main/java/com/auction/entity/AuctionResult.java
@@ -2,6 +2,7 @@ package com.auction.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,12 +11,12 @@ public class AuctionResult {
     private Long sessionId;
     private Long itemId;
     private Long winnerUserId;
-    private Long finalPrice; // 分
+    private BigDecimal finalPrice; // 元
     private Long highestBidId;
     private Integer resultStatus; // 0=流拍 1=成交 2=撤拍
     private Long orderId;
-    private Long commissionFee; // 分
-    private Long depositUsed;   // 分
+    private BigDecimal commissionFee; // 元
+    private BigDecimal depositUsed;   // 元
     private Integer settleStatus; // 0 未结算 1 已结算
     private LocalDateTime settleTime;
     private String remark;
